Refactor: streams used in BuildingManager wherever i could manage. Services package renamed to managers.
