+++ src/main/java/managers/IBroodWarManager.java
@@ -1,4 +1,4 @@
-package services;
+package managers;
 
 public interface IBroodWarManager {
     void manage();
